PyScript housdorf's difference for same-microfaces

+++ housdorff_distance.py
@@ -61,6 +61,7 @@ parser.add_argument("--max-edge", type=float, help="Parameter 4 (max-edge)")
 parser.add_argument("--base-mesh", help="Parameter 5 (base-mesh)")
 parser.add_argument("--target-mesh", help="Parameter 6 (target)")
 parser.add_argument("--metric", default="same-target-edges", help="Parameter 7 (metric)")
+parser.add_argument("--clean", help="Parameter 8 (clean flag)", action="store_true")
 
 args = parser.parse_args()
 
@@ -73,8 +74,9 @@ target_mesh_path = "./Models/" + (args.target_mesh if args.target_mesh else "pal
 exe_path = os.path.join(os.getcwd(), "Debug", "master_thesis.exe")
 
 # Clean the samples dirs
-delete_files_in_directory(directory_path_micro)
-delete_files_in_directory(directory_path_aniso)
+if args.clean:
+  delete_files_in_directory(directory_path_micro)
+  delete_files_in_directory(directory_path_aniso)
 
 params = []
 
@@ -96,12 +98,12 @@ if args.target_mesh is not None:
 if args.metric is not None:
   params.append(f"--metric={args.metric}")
 
-try:
-  subprocess.check_call([exe_path, "build-samples"] + params)
-except subprocess.CalledProcessError as e:
-  print(f"Error while executing the command: {e}")
-except FileNotFoundError as e:
-  print(f"File not found: {e}")
+# try:
+#   subprocess.check_call([exe_path, "build-samples"] + params)
+# except subprocess.CalledProcessError as e:
+#   print(f"Error while executing the command: {e}")
+# except FileNotFoundError as e:
+#   print(f"File not found: {e}")
 
 
 if args.metric == "same-target-edges":
@@ -143,8 +145,39 @@ if args.metric == "same-target-edges":
   output_file.close()
 
 elif args.metric == "same-microfaces":
+  file_list = glob.glob(os.path.join(directory_path_micro, "*"))
+
   with open(f"./Output/Evaluation/{ev_metric_dir}/hausdorff_micro.txt", "w") as output_file:
-    output_file.write("microfaces RMS diag_mesh_0 diag_mesh_1 max mean min n_samples\n")
+    output_file.write("microFaces RMS diag_mesh_0 diag_mesh_1 max mean min n_samples\n")
+
+    for displaced_mesh_path in file_list:
+      ms = pymeshlab.MeshSet()
+      ms.load_new_mesh(target_mesh_path)
+      ms.load_new_mesh(displaced_mesh_path)
+      output_file.write(str(extract_last_decimal_number_from_string(displaced_mesh_path)) + " ")
+
+      for key, value in ms.get_hausdorff_distance().items():
+        output_file.write(str(value) + " ")
+
+      output_file.write("\n")
 
   print(f"Hausdorff's distances > ./Output/Evaluation/{ev_metric_dir}/hausdorff_micro.txt")
+
+  file_list = glob.glob(os.path.join(directory_path_aniso, "*"))
+
+  with open(f"./Output/Evaluation/{ev_metric_dir}/hausdorff_aniso.txt", "w") as output_file:
+    output_file.write("microFaces RMS diag_mesh_0 diag_mesh_1 max mean min n_samples\n")
+
+    for displaced_mesh_path in file_list:
+      ms = pymeshlab.MeshSet()
+      ms.load_new_mesh(target_mesh_path)
+      ms.load_new_mesh(displaced_mesh_path)
+      output_file.write(str(extract_last_decimal_number_from_string(displaced_mesh_path)) + " ")
+
+      for key, value in ms.get_hausdorff_distance().items():
+        output_file.write(str(value) + " ")
+
+      output_file.write("\n")
+  print(f"Hausdorff's distances > ./Output/Evaluation/{ev_metric_dir}/hausdorff_aniso.txt")
+
   output_file.close()
